requirements/requirements.txt
Django
django-environ
django-filter
djangorestframework
djangorestframework-jwt
djangorestframework-filters
mysqlclient
openpyxl
